+++ src/pages/Asset.vue
@@ -1,11 +1,11 @@
 <script lang="ts">
 import { defineComponent, ref, computed } from 'vue';
-import { atomic_api, atomic_market_api } from 'src/api/atomic_assets';
-import { IAsset } from 'atomicassets/build/API/Explorer/Objects';
+import { atomic_market_api } from 'src/api/atomic_assets';
 import {
   ISale,
   IBuyoffer,
-  IMarketOffer
+  IMarketOffer,
+  IMarketAsset
 } from 'atomicmarket/build/API/Explorer/Objects';
 import AssetViewer from 'src/components/atomicAssets/AssetViewer.vue';
 import { useStore } from 'src/store';
@@ -16,7 +16,9 @@ export default defineComponent({
   name: 'Asset',
   components: { AssetViewer },
   setup() {
-    const assetData = ref<IAsset>(new Object({ data: { img: '' } }) as IAsset);
+    const assetData = ref<IMarketAsset>(
+      new Object({ data: { img: '' } }) as IMarketAsset
+    );
     const saleData = ref<ISale>(new Object({}) as ISale);
     const buyofferData = ref<IBuyoffer>(new Object({}) as IBuyoffer);
     const offerData = ref<IMarketOffer>(new Object({}) as IMarketOffer);
@@ -50,7 +52,7 @@ export default defineComponent({
   },
   methods: {
     async getAssetData() {
-      this.assetData = await atomic_api.getAsset(
+      this.assetData = await atomic_market_api.getAsset(
         this.$route.params.asset as string
       );
       console.log(this.assetData);
